1) Targem-style formatting (partially)

+++ CubeV4/DDirect3DTools.h
@@ -7,4 +7,4 @@
 DID3DBlobPtr DCompileShaderFromSourceBlob( DID3DBlobPtr sourceBlob, const char *entryPoint, const char *target, HRESULT *returnCode );
 
 //
-bool DCheckDisplayMode( IDXGIAdapter2 *adapter, u16 width, u16 height, DXGI_RATIONAL refreshRate, HRESULT *returnCode );
+bool DCheckDisplayMode( IDXGIAdapter2 *adapter, u32 width, u32 height, DXGI_RATIONAL refreshRate, HRESULT *returnCode );
