+++ CubeV4/DRenderResourceManager.h
@@ -26,12 +26,11 @@ public:
 	DID3DBlobPtr					CreateBlob( u32 blobSize, HRESULT* returnCode );
 	DVertexBufferViewPtr			CreateVertexBufferView( DID3D11BufferPtr vertices, u32 stride, u32 offset );
 	DIndexBufferViewPtr				CreateIndexBufferView( DID3D11BufferPtr indexBufferPtr, DXGI_FORMAT format, u32 offset );
-	DSimpleMeshPtr					CreateSimpleMesh( DVertexBufferView *vertexView, DIndexBufferView *indexView, DMaterialController::DMaterial *material, u32 indexCount, u32 startVertexLocation, u32 startIndexLocation );
-	DID3D11Texture2DPtr				CreateTexture2D( const DPixelColor *pixels, u16 width, u16 height, HRESULT *returnCode );
-	DID3D11ShaderResourceViewPtr	CreateTexture2DView( ID3D11Texture2D *texture, HRESULT *returnCode );
-	DID3D11InputLayoutPtr			CreateInputLayoutFromVertexShaderBlob( ID3DBlob *vertexShaderBlob, const D3D11_INPUT_ELEMENT_DESC *layout, u16 numLayouts, HRESULT *returnCode );
-	DID3D11SamplerStatePtr			CreateSampler( HRESULT *returnCode );
-
+	DSimpleMeshPtr					CreateSimpleMesh( DVertexBufferView* vertexView, DIndexBufferView* indexView, DMaterialController::DMaterial* material, u32 indexCount, u32 startVertexLocation, u32 startIndexLocation );
+	DID3D11Texture2DPtr				CreateTexture2D( const DPixelColor* pixels, u32 width, u32 height, HRESULT* returnCode );
+	DID3D11ShaderResourceViewPtr	CreateTexture2DView( ID3D11Texture2D* texture, HRESULT* returnCode );
+	DID3D11InputLayoutPtr			CreateInputLayoutFromVertexShaderBlob( ID3DBlob* vertexShaderBlob, const D3D11_INPUT_ELEMENT_DESC* layout, u16 numLayouts, HRESULT* returnCode );
+	DID3D11SamplerStatePtr			CreateSampler( HRESULT* returnCode );
 
 	//
 	DID3D11DevicePtr				GetDevice() const {
